+++ front/src/views/stadistic/Graphics/LinesChart.jsx
