+++ front/src/views/stadistic/Graphics/tripsPerMonth.jsx
@@ -1,4 +1,4 @@
-import { Bar } from 'react-chartjs-2';
+import { Line } from 'react-chartjs-2';
 import {
     Chart as ChartJS,
     CategoryScale,
@@ -23,36 +23,47 @@ ChartJS.register(
 );
 
 
-import {getAllConductores} from '../../redux/actions/index'
+import {getTrips, getCanceledTrips} from '../../../redux/actions/index'
 import { useDispatch, useSelector } from 'react-redux';
 import { useEffect } from 'react';
 
 
-export default function TripsPerDriver() {
+
+export default function TripsPerMonths() {
     const dispatch = useDispatch();
   
     useEffect(async () => {
-      await dispatch(getAllConductores());
+      await dispatch(getTrips());
+    //   await dispatch(getCanceledTrips());
     }, [dispatch]);
+
+    const nombreMeses = [
+        "Ene", "Feb", "Mar", "Abr", "May", "Jun",
+        "Jul", "Ago", "Sep", "Oct", "Nov", "Dic"
+      ];
   
-    const conductores = useSelector((state) => state.allData);
-    const conductoresConTrips = conductores.filter((conductor) => conductor.Trips && conductor.Trips.length > 0);
-  
-    const tripsPorConductor = {};
-  
-    conductoresConTrips.forEach((conductor) => {
-      tripsPorConductor[conductor.name] = conductor.Trips.length;
+    // const viajes = useSelector((state)=>state.viajesCompletados)
+    const viajes = useSelector((state) => state.getTrips);
+    const tripsPorMes = {};
+
+    viajes.forEach((trip) => {
+      const mes = new Date(trip.date).getMonth();
+      if (!tripsPorMes[mes]) {
+        tripsPorMes[mes] = 1; 
+      } else {
+        tripsPorMes[mes] += 1; 
+      }
     });
   
-    const conductoresNombres = Object.keys(tripsPorConductor);
-    const cantidadTrips = Object.values(tripsPorConductor);
+    const meses = Object.keys(tripsPorMes).map((mes) => nombreMeses[parseInt(mes, 10)]);
+    const trips = Object.values(tripsPorMes);
   
     var midata = {
-      labels: conductoresNombres,
+      labels: meses,
       datasets: [
         {
-          label: 'Nro de Viajes',
-          data: cantidadTrips,
+          label: 'Nro Viajes',
+          data: trips,
           tension: 0.5,
           fill: true,
           borderColor: 'rgb(255, 99, 132)',
@@ -68,7 +79,7 @@ export default function TripsPerDriver() {
       scales: {
         y: {
           min: 0,
-          stepSize: 1,
+          stepSize: 1, 
         },
         x: {
           ticks: { color: 'rgb(255, 99, 132)' },
@@ -76,5 +87,5 @@ export default function TripsPerDriver() {
       },
     };
   
-    return <Bar data={midata} options={misoptions} />;
+    return <Line data={midata} options={misoptions} />;
   }
